Update pyscf2qwalk.py
I noticed when trying to convert calculation that used an uncontracted basis that had 10 or more primitives in a given basis type, that the print_orb_coeff would crash when looking up the aosym. This is because the find_label function looked at the second character in the string, which for 10s would be 0 instead of s, which was being used to define the basis label. This change fixes that, by using lstrip to strip all leading integers in the sph_label